fix: move MAX_INVESTIGATION_LOOPS to app/investigation_constants.py
Avoids circular import: app.pipeline.__init__ → graph → app.nodes →
diagnosis node → app.pipeline.constants would fail because app.nodes
is still partially initialized when graph.py tries to import it.

Placing the constant in app/investigation_constants.py (outside both
packages) breaks the cycle. Also fixes import ordering (ruff I001).
