--- app/pipeline/constants.py
@@ -1,5 +1,0 @@
-"""Pipeline constants shared across nodes and routing."""
-
-from __future__ import annotations
-
-MAX_INVESTIGATION_LOOPS = 4
