+++ app/pipeline/routing.py
@@ -4,8 +4,8 @@ from __future__ import annotations
 
 import logging
 
+from app.investigation_constants import MAX_INVESTIGATION_LOOPS
 from app.output import debug_print
-from app.pipeline.constants import MAX_INVESTIGATION_LOOPS
 from app.state import AgentState, InvestigationState
 
 logger = logging.getLogger(__name__)
